+++ src/main/java/com/github/ratel/services/impl/CategoryServiceImpl.java
@@ -1,15 +1,18 @@
 package com.github.ratel.services.impl;
 
 import com.github.ratel.entity.Category;
+import com.github.ratel.exceptions.AppException;
 import com.github.ratel.exceptions.EntityNotFoundException;
 import com.github.ratel.repositories.CategoryRepository;
 import com.github.ratel.services.CategoryService;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -20,6 +23,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional(readOnly = true)
     public List<Category> findAllCategory(Sort sort) {
+        if (Objects.isNull(sort)) {
+            throw new AppException("Invalid parameters value: sort(%)", sort);
+        }
         return this.categoryRepository.findAllByRemovedFalse(sort);
     }
 
@@ -31,14 +37,20 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional(readOnly = true)
-    public Category findById(long id) {
+    public Category findById(Long id) {
+        if (Objects.isNull(id)) {
+            throw new AppException("Invalid parameters value: id(%s)", id);
+        }
         return this.categoryRepository.findByIdAndRemovedFalse(id)
                 .orElseThrow(() -> new EntityNotFoundException("Category not found"));
     }
 
     @Override
     @Transactional(readOnly = true)
-    public Category getByIdForAdmin(long id) {
+    public Category getByIdForAdmin(Long id) {
+        if (Objects.isNull(id)) {
+            throw new AppException("Invalid parameters value: id(%s)", id);
+        }
         return this.categoryRepository.findByIdForAdmin(id)
                 .orElseThrow(() -> new EntityNotFoundException("Category not found"));
     }
@@ -46,12 +58,18 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public Category createCategory(String name) {
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new AppException("Invalid parameters value: name(%s)", name);
+        }
         return this.categoryRepository.save(new Category(name));
     }
 
     @Override
     @Transactional
-    public Category updateCategory(long id, String name) {
+    public Category updateCategory(Long id, String name) {
+        if (ObjectUtils.anyNull(id, name)) {
+            throw new AppException("Invalid parameters value: id(%s), name(%s)", id, name);
+        }
         Category category = this.findById(id);
         category.setName(name);
         return this.categoryRepository.save(category);
@@ -59,7 +77,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional
-    public void deleteCategoryById(long id) {
+    public void deleteCategoryById(Long id) {
+        if (Objects.isNull(id)) {
+            throw new AppException("Invalid parameters value: id(%s)", id);
+        }
         this.categoryRepository.deleteById(id);
     }
