+++ src/main/java/com/github/ratel/services/impl/UserServiceImpl.java
@@ -6,6 +6,7 @@ import com.github.ratel.entity.FileEntity;
 import com.github.ratel.entity.User;
 import com.github.ratel.entity.enums.Roles;
 import com.github.ratel.entity.enums.UserVerificationStatus;
+import com.github.ratel.exceptions.AppException;
 import com.github.ratel.exceptions.ConfirmPasswordException;
 import com.github.ratel.exceptions.EntityNotFoundException;
 import com.github.ratel.exceptions.statuscode.StatusCode;
@@ -59,18 +60,27 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User getCurrentUser(Principal principal) {
+        if (Objects.isNull(principal)) {
+            throw new AppException("Invalid parameters value: principal(%s)", principal);
+        }
         return this.findUserByEmail(principal.getName());
     }
 
     @Override
     @Transactional(readOnly = true)
     public Page<User> findAllUsers(Pageable pageable) {
+        if (Objects.isNull(pageable)) {
+            throw new AppException("Invalid parameters value: pageable(%s)", pageable);
+        }
         return this.userRepository.findAllByRemovedFalse(pageable);
     }
 
     @Override
     @Transactional(readOnly = true)
     public Page<User> findAllUsersForAdmin(Pageable pageable) {
+        if (Objects.isNull(pageable)) {
+            throw new AppException("Invalid parameters value: pageable(%s)", pageable);
+        }
         return this.userRepository.findAllByRemovedTrue(pageable);
     }
 
@@ -101,6 +111,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User findById(Long id) {
+        if (Objects.isNull(id)) {
+            throw new AppException("Invalid parameters value: id(%s)", id);
+        }
         return this.userRepository.findByIdAndRemovedFalse(id)
                 .orElseThrow(() -> new EntityNotFoundException(StatusCode.NOT_FOUND));
     }
@@ -108,6 +121,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User findUserForAdmin(Long userId) {
+        if (Objects.isNull(userId)) {
+            throw new AppException("Invalid parameters value: userId(%s)", userId);
+        }
         return this.userRepository.findUserByIdForAdmin(userId).orElseThrow(
                 () -> new EntityNotFoundException(StatusCode.NOT_FOUND)
         );
@@ -116,12 +132,18 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User checkUserByEmail(String email) {
+        if (Objects.isNull(email)) {
+            throw new AppException("Invalid parameters value: email(%s)", email);
+        }
         return this.userRepository.findByEmailAndRemovedFalse(email).orElse(null);
     }
 
     @Override
     @Transactional(readOnly = true)
     public boolean findUserByRole(Roles role) {
+        if (Objects.isNull(role)) {
+            throw new AppException("Invalid parameters value: role(%s)", role);
+        }
         User user = this.userRepository.findUserByRolesAndRemovedFalse(role).orElse(null);
         return user != null;
     }
@@ -129,6 +151,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User findUserByEmail(String email) {
+        if (Objects.isNull(email)) {
+            throw new AppException("Invalid parameters value: email(%s)", email);
+        }
         return userRepository.findByEmailAndRemovedFalse(email)
                 .orElseThrow(() -> new EntityNotFoundException(StatusCode.NOT_FOUND));
     }
@@ -136,6 +161,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public User findByEmailAndPassword(String email, String password) {
+        if (ObjectUtils.anyNull(email, password)) {
+            throw new AppException("Invalid parameters value: email or password");
+        }
         User user = this.findUserByEmail(email);
         if (passwordEncoder.matches(password, user.getPassword())) {
             return user;
@@ -146,6 +174,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public User createUser(Roles role, CreateUserRequest payload) {
+        if (ObjectUtils.anyNull(role, payload)) {
+            throw new AppException("Invalid parameters value: role(%s) or CreateUserRequest", role);
+        }
         CheckUtil.checkUserByEmail(this.checkUserByEmail(payload.getEmail()));
         CheckUtil.checkPassAndConfirmPass(payload.getPassword(), payload.getConfirmPassword());
         Address address = new Address();
@@ -169,6 +200,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public User editUser(UserUpdateRequest updateRequest, MultipartFile image) {
+        if (ObjectUtils.anyNull(image, updateRequest)) {
+            throw new AppException("Invalid parameters value: image(%s) or UserUpdateRequest", image);
+        }
         User user = this.findById(updateRequest.getId());
         UserTransferObj.updateUser(user, updateRequest);
         FileEntity fileEntity = null;
@@ -183,12 +217,18 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public User updateUser(User user) {
+        if (Objects.isNull(user)) {
+            throw new AppException("Invalid parameters value: user(%s)", user);
+        }
         return this.userRepository.save(user);
     }
 
     @Override
     @Transactional
     public void deleteUserById(Long userId) {
+        if (Objects.isNull(userId)) {
+            throw new AppException("Invalid parameters value: userId(%s)", userId);
+        }
         this.cartService.deleteCartByUserId(userId);
         this.userRepository.deleteById(userId);
     }
